feat: move logic creating town from client to server
