+++ package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "cd public/src/ && nodemon server.js"
+    "start": "cd server && nodemon server.js"
   },
   "repository": {
     "type": "git",
