+++ server/util.js
@@ -1,0 +1,5 @@
+function getRandomNumber({ min, max }) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export { getRandomNumber };
